+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * 【core  】    修改Dict在非大小写敏感状态下get也不区分大小写（issue#722@Github）
 * 【core  】    StrUtil增加contains方法（issue#716@Github）
 * 【core  】    QrCodeUtil增加背景透明支持（pr#89@Gitee）
+* 【core  】    增加农历ChineseDate（pr#90@Gitee）
 
 ### Bug修复
 * 【core 】     修复NumberUtil.mul中null的结果错误问题（issue#I17Y4J@Gitee）
